+++ oop-presentation/src/main/kotlin/com/jaeyong/oop/presentation/health/api/HealthController.kt
@@ -1,6 +1,6 @@
-package com.jaeyong.oop.presentation.api
+package com.jaeyong.oop.presentation.health.api
 
-import com.jaeyong.oop.application.usecase.HealthCheckUseCase
+import com.jaeyong.oop.application.health.usecase.HealthCheckUseCase
 import com.jaeyong.oop.presentation.response.ApiResponse
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
