+++ oop-presentation/src/main/kotlin/com/jaeyong/oop/presentation/user/request/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.jaeyong.oop.presentation.user
+package com.jaeyong.oop.presentation.user.request
 
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
